Use source of symlinked OSS devices
If you set the dev attribute of one of the OSS devices to a symlink in
/dev/disk/ e.g. /dev/disk/by-id/<id>, then run the role with the file
system already mounted, the following task fails:

TASK [roles/stackhpc.beegfs : Attempt to format if the device is not
mounted or if beegfs_force_format is true]

mkfs.xfs: cannot open /dev/disk/by-path/pci-0000:18:00.0-sas-phy4-lun-0:
Device or resource busy

This is because the task runs only if the device is not mounted, but the
symlink source is displayed in the mount list rather than the target.

This commit changes to use the source of the link when checking the list
of mounted devices.

We also introduce a check for the existence of the device, and fail if
it does not exist.

Fixes: #50