Umount meta data device only in case of reformat
Prior to this change, running the playbook a second time will not work
because the meta.yml playbook will try to unmount the meta data device
while the service is still running and also it is unnecessary.
Additionally add explicit filesystem and mount options for the metadata
devices.
Small bugfix for newer ansbile file module and allow the user allow for
meta data device tunings